requirements.txt
# Core Data Science Libraries
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Machine Learning
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0

# Deep Learning
tensorflow==2.13.0
keras==2.13.1

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Web Framework
flask==2.3.2
flask-cors==4.0.0
gunicorn==21.2.0

# MLOps and Experiment Tracking
mlflow==2.5.0
dvc==3.15.0

# Model Optimization
optuna==3.2.0
hyperopt==0.2.7

# Data Validation
great-expectations==0.17.10

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Code Quality
pylint==2.17.4
flake8==6.0.0
black==23.7.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
joblib==1.3.1
tqdm==4.65.0

# API Documentation
flask-swagger-ui==4.11.1

# Logging and Monitoring
loguru==0.7.0

# Data Processing
openpyxl==3.1.2
xlrd==2.0.1

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# HTTP Requests
requests==2.31.0

# Environment Management
python-decouple==3.8
